safari-600.7-branch/LayoutTests/js/string-capitalization-expected.txt

This test checks that toLowerCase and toUpperCase handle certain non-trivial cases correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String("Ağ€").toLowerCase() is "ağ¨"
PASS String("ağ¨").toUpperCase() is "Ağ€"
PASS String("ÎšÎŸÎ£ÎœÎŸÎ£ ÎšÎŸÎ£ÎœÎŸÎ£").toLowerCase() is "ÎºÎ¿ÏƒÎ¼Î¿Ï‚ ÎºÎ¿ÏƒÎ¼Î¿Ï‚"
PASS String("ÃŸ").toUpperCase() is "SS"
PASS String("Å‰").toUpperCase() is "Ê¼N"
PASS String("Ç°").toUpperCase() is "JÌŒ"
PASS String("ï¬ƒ").toUpperCase() is "FFI"
PASS String("FFI").toLowerCase() is "ffi"
PASS String("Ä²").toLowerCase() is "Ä³"
PASS expected[String.fromCharCode(0xA78D).toLowerCase()] is true
PASS expected[String.fromCharCode(0x0265).toUpperCase()] is true
PASS expected[String.fromCharCode(0x10C7).toLowerCase()] is true
PASS expected[String.fromCharCode(0x2D27).toUpperCase()] is true
PASS expected[String.fromCharCode(0x2D2D).toLowerCase()] is true
PASS expected[String.fromCharCode(0x10CD).toUpperCase()] is true
PASS expected[String.fromCharCode(0x2CF2).toLowerCase()] is true
PASS expected[String.fromCharCode(0x2CF3).toUpperCase()] is true
PASS expected[String.fromCharCode(0xA792).toLowerCase()] is true
PASS expected[String.fromCharCode(0xA793).toUpperCase()] is true
PASS expected[String.fromCharCode(0xA7AA).toLowerCase()] is true
PASS expected[String.fromCharCode(0x0266).toUpperCase()] is true
PASS successfullyParsed is true

TEST COMPLETE
